Update GeoRefAdditionalStatisticsData.java
The GeoLite2-City service does not cope well when asked about private adresses. It fails throwing an AddressNotFoundException, which in turn is _not_ catched here(!) (This is another issue this patch does _not_cope with).
This patch uses isSiteLocalAddress() to keep local addresses away from the GeoLite service in the first place.

Sadly, this won't work anymore in dspace-cris-master, as the address is then handed over as string.