Update GeoRefAdditionalStatisticsData.java
The GeoLite2-City service does not cope well when asked about private adresses.
This patch uses InetAddress.isSiteLocalAddress() to keep private addresses away from the GeoLite service in the first place, preventing wasted lookups.
The instatiation of InetAddress was skipped after dspace-6_x_x-cris and current master. If this was a performance issue, skip this patch.