Enhance DOIIdentifierProviderUsingHandles
Use the handle's suffix, combined with the DOI prefix and
namespaceseparator. This aligns the DOIIdentifierProviderUsingHandles
more closely to the DOIIdentifierProvider and ensures that the right DOI
prefix is used. For DOIs and handles being identically, you'll have to
set the handle prefix to the DOI prefix a / and the DOI
namespaceseparator. We would advise to leave the namespace separator
empty.